Split into multiple testing stanzas and modules for backends
Common/Test.hs holds all common tests and functionality for the backends
