+++ esqueleto.cabal
@@ -82,11 +82,45 @@ library
   else
     ghc-options: -Wall
 
-test-suite test
+-- test-suite test-common
+--   type: detailed-0.9
+--   ghc-options:    -Wall
+--   hs-source-dirs: test
+--   test-module:    Common.Test
+--   build-depends:
+--       -- Library dependencies used on the tests.  No need to
+--       -- specify versions since they'll use the same as above.
+--       base, persistent, transformers, resourcet, text
+--
+--       -- Test-only dependencies
+--     , conduit             >= 1.1
+--     , containers
+--     , HUnit
+--     , QuickCheck
+--     , hspec               >= 1.8
+--     , persistent-sqlite   >= 2.1.3
+--     , persistent-template >= 2.1
+--     , monad-control
+--     , monad-logger        >= 0.3
+--     , time                >= 1.5.0.1 && <= 1.8.0.2
+--
+--       -- This library
+--     , esqueleto
+--
+--     , postgresql-simple     >= 0.2
+--     , postgresql-libpq      >= 0.6
+--     , persistent-postgresql >= 2.0
+--
+--     , mysql-simple          >= 0.2.2.3
+--     , mysql                 >= 0.1.1.3
+--     , persistent-mysql      >= 2.0
+
+test-suite postgresql
   type: exitcode-stdio-1.0
   ghc-options:    -Wall
   hs-source-dirs: test
-  main-is:        Test.hs
+  other-modules: Common.Test
+  main-is:        PostgreSQL/Test.hs
   build-depends:
       -- Library dependencies used on the tests.  No need to
       -- specify versions since they'll use the same as above.
@@ -98,10 +132,7 @@ test-suite test
     , HUnit
     , QuickCheck
     , hspec               >= 1.8
-    , persistent-sqlite   >= 2.1.3
-    , persistent-template >= 2.1
     , monad-control
-    , monad-logger        >= 0.3
     , time                >= 1.5.0.1 && <= 1.8.0.2
 
       -- This library
@@ -110,13 +141,65 @@ test-suite test
     , postgresql-simple     >= 0.2
     , postgresql-libpq      >= 0.6
     , persistent-postgresql >= 2.0
+    , persistent-template >= 2.1
+    , monad-control
+    , monad-logger        >= 0.3
+
+
+test-suite mysql
+  type: exitcode-stdio-1.0
+  ghc-options:    -Wall
+  hs-source-dirs: test
+  other-modules:  Common.Test
+  main-is:        MySQL/Test.hs
+  build-depends:
+      -- Library dependencies used on the tests.  No need to
+      -- specify versions since they'll use the same as above.
+      base, persistent, transformers, resourcet, text
+
+      -- Test-only dependencies
+    , conduit             >= 1.1
+    , containers
+    , HUnit
+    , QuickCheck
+    , hspec               >= 1.8
+    , monad-control
+    , time                >= 1.5.0.1 && <= 1.8.0.2
+
+      -- This library
+    , esqueleto
 
     , mysql-simple          >= 0.2.2.3
     , mysql                 >= 0.1.1.3
     , persistent-mysql      >= 2.0
+    , persistent-template >= 2.1
+    , monad-control
+    , monad-logger        >= 0.3
 
-  if flag(postgresql)
-    cpp-options: -DWITH_POSTGRESQL
 
-  if flag(mysql)
-    cpp-options: -DWITH_MYSQL
+test-suite sqlite
+  type: exitcode-stdio-1.0
+  ghc-options:    -Wall
+  hs-source-dirs: test
+  other-modules:  Common.Test
+  main-is:        SQLite/Test.hs
+  build-depends:
+      -- Library dependencies used on the tests.  No need to
+      -- specify versions since they'll use the same as above.
+      base, persistent, transformers, resourcet, text
+
+      -- Test-only dependencies
+    , conduit             >= 1.1
+    , containers
+    , HUnit
+    , QuickCheck
+    , hspec               >= 1.8
+    , monad-control
+    , time                >= 1.5.0.1 && <= 1.8.0.2
+
+      -- This library
+    , esqueleto
+
+    , persistent-sqlite   >= 2.1.3
+    , persistent-template >= 2.1
+    , monad-logger        >= 0.3
